Transformers - base making tests on abstract base class
