+++ Symfony/CS/Tests/Tokenizer/Transformer/DynamicVarBraceTest.php
@@ -12,7 +12,6 @@
 namespace Symfony\CS\Tests\Tokenizer\Transformer;
 
 use Symfony\CS\Tests\Tokenizer\AbstractTransformerTestBase;
-use Symfony\CS\Tokenizer\Tokens;
 
 /**
  * @author Dariusz Rumiński <dariusz.ruminski@gmail.com>
@@ -24,20 +23,7 @@ class DynamicVarBraceTest extends AbstractTransformerTestBase
      */
     public function testProcess($source, array $expectedTokens = array())
     {
-        $tokens = Tokens::fromCode($source);
-
-        $this->assertSame(
-            count($expectedTokens),
-            array_sum(array_map(
-                function ($item) { return count($item); },
-                $tokens->findGivenKind(array_map(function ($name) { return constant($name); }, $expectedTokens))
-            ))
-        );
-
-        foreach ($expectedTokens as $index => $name) {
-            $this->assertSame(constant($name), $tokens[$index]->getId());
-            $this->assertSame($name, $tokens[$index]->getName());
-        }
+        $this->makeTest($source, $expectedTokens);
     }
 
     public function provideProcessCases()
